+++ devtools/clean_dist_check.py
@@ -1,5 +1,6 @@
-"""Clean dist files after tox -e check"""
+"""Clean dist files after tox -e check."""
 from pathlib import Path
+
 
 folder = Path(__file__).resolve().parents[1]